+++ lib/src/models/ble_property.dart
@@ -2,17 +2,9 @@ enum BleInputProperty {
   disabled,
   notification,
   indication;
-
-  const BleInputProperty();
-
-  factory BleInputProperty.parse(int index) => BleInputProperty.values[index];
 }
 
 enum BleOutputProperty {
   withResponse,
   withoutResponse;
-
-  const BleOutputProperty();
-
-  factory BleOutputProperty.parse(int index) => BleOutputProperty.values[index];
 }