+++ lib/src/universal_ble_linux/universal_ble_linux.dart
@@ -3,9 +3,9 @@ import 'dart:async';
 import 'package:bluez/bluez.dart';
 import 'package:flutter/services.dart';
 import 'package:universal_ble/src/models/model_exports.dart';
-import 'package:universal_ble/src/universal_ble_filter_util.dart';
+import 'package:universal_ble/src/utils/universal_ble_filter_util.dart';
 import 'package:universal_ble/src/universal_ble_platform_interface.dart';
-import 'package:universal_ble/src/universal_logger.dart';
+import 'package:universal_ble/src/utils/universal_logger.dart';
 
 class UniversalBleLinux extends UniversalBlePlatform {
   UniversalBleLinux._();
@@ -185,14 +185,23 @@ class UniversalBleLinux extends UniversalBlePlatform {
 
     List<BleService> services = [];
     for (final service in device.gattServices) {
+      String serviceId = service.uuid.toString();
+
       final characteristics = service.characteristics.map((e) {
         final properties = List<CharacteristicProperty>.from(e.flags
             .map((e) => e.toCharacteristicProperty())
             .where((element) => element != null)
             .toList());
-        return BleCharacteristic(e.uuid.toString(), properties);
+        return BleCharacteristic.withMetaData(
+          deviceId: deviceId,
+          serviceId: serviceId,
+          uuid: e.uuid.toString(),
+          properties: properties,
+        );
       }).toList();
-      services.add(BleService(service.uuid.toString(), characteristics));
+      services.add(
+        BleService(serviceId, characteristics),
+      );
     }
     return services;
   }
@@ -604,7 +613,7 @@ extension BlueZDeviceExtension on BlueZDevice {
     return BleDevice(
       name: name,
       deviceId: address,
-      isPaired: paired,
+      paired: paired,
       rssi: rssi,
       isSystemDevice: isSystemDevice,
       services: uuids.map((e) => e.toString()).toList(),