+++ test/universal_ble_stream_controller_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:universal_ble/src/universal_ble_stream_controller.dart';
+import 'package:universal_ble/src/utils/universal_ble_stream_controller.dart';
 
 void main() {
   group("Test UniversalBleStreamController", () {