+++ CHANGELOG.md
@@ -1,6 +1,12 @@
-## 0.20
-* Fix `getBluetoothAvailabilityState` reporting wrong Bluetooth status on browsers where Web Bluetooth can be globally disabled
-* `startScan` API on Web platforms now throws a `WebBluetoothGloballyDisabled` exception if scanning cannot proceed due to Web Bluetooth being globally disabled
+## 0.20.0
+* Add new high level API. Services are auto-discovered. the BleDevice class offers convenient member methods and properties.
+* BREAKING CHANGE: `bleDevice.isPaired` is now `bleDevice.paired`
+* Deprecate `readValue` in favor of `read`
+* Deprecate `writeValue` in favor of `write`
+* Deprecate `setNotifiable` in favor of `subscribeNotifications`, `subscribeIndications`, `unsubscribe`
+* Fix `getBluetoothAvailabilityState` reporting wrong Bluetooth status on browsers where Web Bluetooth can globally be disabled
+* `startScan` on web now throws a `WebBluetoothGloballyDisabled` exception if scanning cannot proceed due to Web Bluetooth being globally disabled
+* EXAMPLE APP: Migrate to the new high level API
 
 ## 0.19.0
 * Get and prefer advertised name of scanned devices on Apple